#2-T3-Create Wallet from request done
